Note:
My commit b0f733b07678c47cc78eeaba5463eca933bfa6f9 stated it was
cleaning up the HF code. I actually inadvertantly fixed a bug
in the code as the Vh and F were not being mixed properly, and
now they are. This only seemed to effect the correlated molecule
(as it is so sensitive to energies), other systems have been rerun
and the results are the same.

This commit just contains a added comment showing where
the relevent change is.