fixed bug in full preconditioner
Was incorrectly using transpose of preconditioning matrix
for density mixing.
This is now fixed and there is a unit test that goes in this direction.

Unfortunately, the fix didn't improve the convergence at all...
full preconditionin is still worse than no preconditioning for
the correlated test system with Pulay turned on(18 instead of 14 iterations).
The step-system still does not converge.