added xc contribution to preconditioner
Finally, using the preconditioner is *better* than not using it!
For the correlated system, we now have 10 iterations with full
preconditioner vs. 14 iterations without.

It seems to me that in our systems the balance between Hartree and
xc contribution is tilted towards the xc side of things
(this would also explain why the Kerker preconditioner does not help).

Unfortunately, the step-system still does not converge...