hdf5 now supports compression + re-writing files
For the default parameters file with grid spacing replaced by 501,
the size of the hdf5 file is 2.6MB compared to 14M for the pickle
files combined (main contribution were the eigfs).

As far as overwriting datasets in hdf5 files is concerned:
Beware that this is not effective and results in size increases
of the hdf5 file.