fixed bug in pickling parameters file
When pickling the parameters file in a run from a different folder,
it complained that it wasn't able to pickle nested functions
(which pickle shouldn't be able to do anyways).

The solution for the moment is to remove the nested functions
(v_ext, v_pert, v_pert_im) from the parameters object before
it's pickled.