Changed initial wavefunction in EXT codes.
The initial wavefunction in EXT2 and EXT3 can now be constructed
from the HF, LDA or NON orbitals (HF and LDA must be provided, NON
is the default and will be calculated if not provided). Alternatively,
an EXT wavefunction can be given e.g. the one from a previous run. The
parameters file and input file have been updated to reflect these changes.
Bug fixes in EXT3. EXT1 now saves the wavefunction to file if needed.