Initial psi in EXT, order of codes, saving results
A bad starting guess to the initial wavefunction in EXT e.g.
wrong number of electrons per potential well, can mean that
the true ground-state is not found due to numerical issues.
You can now select 'qho' as the initial starting guess which
will ensure stable (but longer) convergence to the true
ground-state if a reliable starting guess can not be provided.
Also, orbitals for LDA, NON and HF are now saved by default.
Also, the order of the codes has changed: EXT will run after the
various approximations. This makes selecting these approximations
as guesses to the initial wavefunction in EXT easier.