Changed default initial wavefunction in EXT2/3.
A very bad initial wavefunction can end up giving the wrong
ground-state e.g. if you start with the wrong number of electrons
per well. The default is now the quantum harmonic oscillator 'qho',
which will ensure stable convergence to the true ground-state. If
a reasonable approximation can be provided, use that to get a big
speed up in the imaginary time propagation.