Fixed bug in current density.
Fixed a bug in the current density which wrongly shifted the true
current density to the left by one grid point. Rewrote RE_Utilities.
Fixed a small inconsistency in the number of elements in time-dependent
arrays (now imax, instead of imax+1).

STILL TO FIX: BUG IN CURRENT DENSITY WHEN USING IMAGINARY POTENTIALS.