printing function refreshes only every x sec
The high frequency of printouts is a problem in jupyter notebooks,
especially if the jupyter server is located on a remote machine.

I've now changed the sprint function such that when overwriting lines
(such as convergence checking), it only updates once a given time
interval is exceeded (by default: 50ms).