Speed up in EXT2/3 and other modifications
Massive speed up in real time propagation in EXT2 and EXT3 by
modifying the way the wavefunction is passed from a 1D array
to a 2D/3D array when calculating the density. This change is
also made in OPT. Removed trapezium rule calculations from all
codes (OPT and RE) due to it being more expensive, but no more
accurate than the rectangle rule for our calculations with a
constant dx.