Changes to EXT code and Cython modules (see below)
1. For long EXT runs (particularly when calculating
excited states), a large proportion of the computational
time was accessing the log file. This has been mitigated
by adding a new argument to the sprint function in
input.py. The EXT2/3 codes now only accesses the log
file at certain time steps (every 1000 in imaginary
time and every 100 in real time).

2. EXT_cython and RE_cython now handle numpy arrays
in a different way.

3. Removed two defunct functions from EXT_cython.

4. Changed a few strings and tidied up a bit in
EXT2/3 codes (reflected in the OPT code when the
imaginary time function is called).