Removing atol parameter
Future scipy releases will change the default atol parameter in
scipy.sparse.linalg.cg(), and so we added atol='legacy' to our
function calls, so that we keep the old default value. However,
Research2 has an old version of scipy which does not support this,
so for now we have removed it.